% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max_scale.R
\name{min_max_scale}
\alias{min_max_scale}
\title{Find the difference to the minimum value, with respect to the min-max difference}
\usage{
min_max_scale(x)
}
\arguments{
\item{x}{A list of numbers}
}
\value{
A scale with respect to the the min-max difference
}
\description{
Find the difference to the minimum value, with respect to the min-max difference
}
\examples{
min_max_scale(1:3)
}
