% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_academic.R
\name{theme_academic}
\alias{theme_academic}
\title{Adjust the layout of a plot}
\usage{
theme_academic(base_size = 12)
}
\arguments{
\item{base_size}{size of the plot}
}
\value{
a plot with the input base size, a bold plot title in the center and no grid
}
\description{
Adjust the layout of a plot
}
\examples{
library(ggplot2)
p<- ggplot(data = mtcars) + geom_point(aes(x= mpg, y= hp)) + labs(title = "plot_title")
p
p + theme_academic()
}
